Generalization of ArrayPlotter
